api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==3.2.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.4
Deprecated==1.3.1
ecdsa==0.19.1
fastapi==0.115.8
h11==0.14.0
idna==3.10
limits==5.6.0
packaging==26.0
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.5.0
python-multipart==0.0.22
requests==2.32.5
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.6.3
uvicorn==0.34.0
wrapt==2.1.1
# Missing dependencies added back
SQLAlchemy
psycopg[binary]
prometheus-client
aiofiles
redis
# ML — Threat Detection
scikit-learn>=1.4.0
joblib>=1.3.0
numpy>=1.26.0
pandas>=2.2.0

# HTTP client — Threat Intelligence (AbuseIPDB)
httpx>=0.27.0

# Testing
pytest
httpx
